Move settings-data.js into the main extension bundle
In Manifest V3 the background code runs in a Service Worker rather than
a web page. This means that there is no `window` global and the entry
point needs to be a single JavaScript file.

Address these requirements by building the generated settings-data.js
file into the main extension bundle.