Initialize tab states on startup by checking for presence of client
Change how the Hypothesis-related state for each tab is initialized on
startup. Instead of caching the previous state in localStorage, instead
run code in each tab to query whether the client is currently loaded and
then mark the tab as active or inactive accordingly.

This change solves several problems:

 - It fixes an issue where the extension failed to inject into existing
   tabs after an installation, until the tab was navigated, because the
   `ready` state for these tabs was incorrectly initialized to `false`.
   This means that the extension incorrectly treated pre-existing tabs
   as being in a loading state.

 - It avoids the possibility of the saved state in the extension's
   localStorage getting out of sync with the actual state. This can
   happen if the tabs in the browser change at all while the extension's
   background page is not running.

 - It means that `TabStore`, and its use of localStorage, are no longer
   needed. localStorage is not available in Manifest V3 when the
   background page becomes a Service Worker, so we would have had to
   replace this with some async storage mechanism.

A downside to this approach is that code has to be run in each tab to determine
the client state. This would not work if the client lost permission to run code
in arbitrary tabs (eg. if browsers made it possible to revoke this required
permission in future). In that case it would have to assume that the client was
not active in the tab. Checking whether the client is active in a tab is done
with a simple CSS selector query, so we might be able to use some declarative
API as a replacement in that case.